--- a/src/formats/MMTF.cpp
+++ b/src/formats/MMTF.cpp
@@ -24,6 +24,7 @@
 #include "chemfiles/Connectivity.hpp"
 
 #include "chemfiles/types.hpp"
+#include "chemfiles/parse.hpp"
 #include "chemfiles/warnings.hpp"
 #include "chemfiles/ErrorFmt.hpp"
 #include "chemfiles/external/optional.hpp"