--- a/src/formats/PDB.cpp
+++ b/src/formats/PDB.cpp
@@ -28,6 +28,7 @@
 
 #include "chemfiles/types.hpp"
 #include "chemfiles/utils.hpp"
+#include "chemfiles/parse.hpp"
 #include "chemfiles/warnings.hpp"
 #include "chemfiles/ErrorFmt.hpp"
 #include "chemfiles/external/optional.hpp"