This PR improves the performance of text format by around 20%. Using [hyperfine](https://github.com/sharkdp/hyperfine) on `xyz` and `lammps-data` tests, i get

**before**
```
$ hyperfine --warmup 3 ./xyz ./lammps-data
Benchmark #1: ./xyz
  Time (mean ± σ):      72.1 ms ±   2.0 ms    [User: 64.2 ms, System: 4.9 ms]
  Range (min … max):    69.5 ms …  77.3 ms    40 runs

Benchmark #2: ./lammps-data
  Time (mean ± σ):      92.0 ms ±   2.7 ms    [User: 82.9 ms, System: 5.6 ms]
  Range (min … max):    87.2 ms …  98.2 ms    31 runs
```

**After**
```
$ hyperfine --warmup 3 ./xyz ./lammps-data
Benchmark #1: ./xyz
  Time (mean ± σ):      56.2 ms ±   1.5 ms    [User: 48.5 ms, System: 4.6 ms]
  Range (min … max):    53.4 ms …  60.4 ms    51 runs

Benchmark #2: ./lammps-data
  Time (mean ± σ):      74.8 ms ±   1.9 ms    [User: 65.8 ms, System: 5.2 ms]
  Range (min … max):    72.0 ms …  79.2 ms    38 runs
```


-----

The first part of the PR introduces a C++11 implementation of C++17 `std::string_view` (https://github.com/martinmoene/string-view-lite) in chemfiles, and uses it for `split`/`trim`, removing a lot of small strings allocations.

The second part introduces our own `strtod` and `strtoll`-like functions, which are faster than the standard library ones since they (a) don't care for hexadecimal or octal notations, (b) don't support user-specific locale and (c) don't ensure perfect rounding, especially for denormal floats. The last point mean that we can be off from strtod, usually by a **relative** error of 1e-15, and no more than 1e-14 in my tests.

I have a follow-up PR using string_view in `TextFile`, but this one is already pretty big, so let's start with this!