@Luthaf 2019-09-10
> but I think so debug code made it into the commit.

Whoops, thanks for spotting it!